Moved NavigationView, connectionRow and MainView to separate files
